Removing templates directory and updating readme.
I've removed the templates directory so that Django will look for templates in the static root instead.
This just makes it easy to automatically override the default mezzanine template
with the django-webpack-loader template when you run collectstatic.

I've edited the readme to reflect the changes introduced with the Mezzanine CMS intergration.